+++ Changelog.md
@@ -1,6 +1,8 @@
 ## Gazebo 5.0
 
 ### Gazebo 5.1.0
+1. Add Link::GetWorldAngularMomentum function and test.
+
 1. Implement Coulomb joint friction for DART
   * [Pull request #1427](https://bitbucket.org/osrf/gazebo/pull-request/1427)
   * [Issue #1281](https://bitbucket.org/osrf/gazebo/issue/1281)